Split query param related tests into smaller modules
